Properly split patterns for MIDI finally
