+++ MIDIProgramSplitter/Program.cs
@@ -17,15 +17,18 @@ internal static class Program
 		args = new string[]
 		{
 			//@"E:\Music\MIDIs\Pokemon B2W2\Music\Roaming Legendary Battle format1.mid",
-			@"E:\Music\MIDIs\Pokemon B2W2\Music\B2W2KantoChampion format1.mid",
+			//@"E:\Music\MIDIs\Pokemon B2W2\Music\B2W2KantoChampion format1.mid",
+			//@"E:\Music\MIDIs\Pokemon B2W2\Music\Elite4Battle format1.mid",
+			//@"E:\Music\MIDIs\Pokemon B2W2\Music\Iris Battle format1.mid",
 			//@"E:\Music\MIDIs\Pokemon B2W2\Music\Colress Battle.mid",
-			//@"E:\Music\MIDIs\DS Rips\HGSS\BATTLE1\SEQ_GS_VS_TRAINER format1.mid",
+			@"E:\Music\MIDIs\DS Rips\HGSS\BATTLE1\SEQ_GS_VS_TRAINER format1.mid",
 			//@"E:\Music\MIDIs\DS Rips\HGSS\BATTLE2\SEQ_GS_VS_RIVAL format1.mid",
 			//@"E:\Music\MIDIs\DS Rips\HGSS\BATTLE6\SEQ_GS_VS_RAIKOU format1.mid",
 			//@"C:\Users\Kermalis\Documents\Development\GitHub\pokeemerald\sound\songs\midi\mus_vs_aqua_magma.mid",
+			//@"C:\Users\Kermalis\Documents\Development\GitHub\pokeemerald\sound\songs\midi\mus_surf.mid",
 
-			@"E:\Music\MIDIs\Split Rips\BWB2W2"
-			//@"E:\Music\MIDIs\Split Rips\HGSS"
+			//@"E:\Music\MIDIs\Split Rips\BWB2W2"
+			@"E:\Music\MIDIs\Split Rips\HGSS"
 			//@"E:\Music\MIDIs\Split Rips\RSE"
 		};
 #endif
@@ -46,13 +49,14 @@ internal static class Program
 			throw new Exception("Output directory not found: \"" + outDir + '\"');
 		}
 		string outFile = Path.Combine(outDir, Path.GetFileName(inFile));
+		const string OUT = @"C:\Users\Kermalis\Documents\Development\GitHub\MIDIProgramSplitter\TestOUT.flp";
 
 		using (FileStream fs = File.OpenRead(inFile))
 		{
 			var inMIDI = new MIDIFile(fs);
 			var splitter = new Splitter(inMIDI);
 			splitter.SaveMIDI(outFile);
-			splitter.Test_WriteFLP();
+			splitter.SaveFLP(OUT);
 		}
 
 		Console.WriteLine();
