+++ MIDIProgramSplitter/Splitter.cs
@@ -1,12 +1,11 @@
 ﻿using Kermalis.MIDI;
-using MIDIProgramSplitter.FLP;
 using System;
 using System.Collections.Generic;
 using System.IO;
 
 namespace MIDIProgramSplitter;
 
-internal sealed class Splitter
+internal sealed partial class Splitter
 {
 	private readonly MIDIFile _inMIDI;
 	private readonly uint _maxTicks;
@@ -88,108 +87,12 @@ internal sealed class Splitter
 
 		foreach (TrackData t in _splitTracks)
 		{
-			t.AddNewTracks(midi);
+			t.MIDI_AddNewTracks(midi);
 		}
 
 		using (FileStream fs = File.Create(outFile))
 		{
 			midi.Save(fs);
 		}
-	}
-
-	public void Test_WriteFLP()
-	{
-		const string OUT = @"C:\Users\Kermalis\Documents\Development\GitHub\MIDIProgramSplitter\TestOUT.flp";
-
-		using (FileStream s = File.Create(OUT))
-		{
-			var w = new FLProjectWriter(ppqn: _inMIDI.HeaderChunk.TimeDivision.PPQN_TicksPerQuarterNote);
-
-			FLP_ReadMetaTrack(w, out decimal tempo, out byte timeSigNum, out byte timeSigDenom);
-			w.CurrentTempo = tempo;
-			w.TimeSigNumerator = timeSigNum;
-			w.TimeSigDenominator = timeSigDenom;
-
-			foreach (TrackData t in _splitTracks)
-			{
-				t.FLP_HandleTrack(w, _maxTicks);
-			}
-
-			w.Write(s);
-		}
-	}
-	private void FLP_ReadMetaTrack(FLProjectWriter w, out decimal tempo, out byte timeSigNum, out byte timeSigDenom)
-	{
-		const int DEFAULT_MIDI_TEMPO = 120;
-
-		tempo = DEFAULT_MIDI_TEMPO;
-		MIDIEvent? firstTempo = null;
-		FLAutomation? tempoAuto = null;
-		timeSigNum = 4;
-		timeSigDenom = 4;
-		MIDIEvent? firstTimeSig = null;
-		bool createdFirstTimeSigMarker = false;
-
-		for (MIDIEvent? e = _metaTrack.First; e is not null; e = e.Next)
-		{
-			var m = (MetaMessage)e.Message;
-			switch (m.Type)
-			{
-				case MetaMessageType.Tempo:
-				{
-					FLP_HandleTempo(e, m, w, ref tempo, ref firstTempo, ref tempoAuto);
-					break;
-				}
-				case MetaMessageType.TimeSignature:
-				{
-					FLP_HandleTimeSig(e, m, w, ref timeSigNum, ref timeSigDenom, ref firstTimeSig, ref createdFirstTimeSigMarker);
-					break;
-				}
-			}
-		}
-
-		tempoAuto?.PadTempoPoints(_maxTicks, DEFAULT_MIDI_TEMPO);
-	}
-	private void FLP_HandleTempo(MIDIEvent e, MetaMessage m, FLProjectWriter w,
-		ref decimal tempo, ref MIDIEvent? firstTempo, ref FLAutomation? tempoAuto)
-	{
-		m.ReadTempoMessage(out _, out decimal bpm);
-		if (firstTempo is null)
-		{
-			tempo = bpm;
-			firstTempo = e;
-			return;
-		}
-
-		// This is the 2nd or after change. 2nd will create the autoclip
-		if (tempoAuto is null)
-		{
-			tempoAuto = new FLAutomation("Tempo", FLAutomation.MyType.Tempo, null);
-			w.Automations.Add(tempoAuto);
-			w.AddToPlaylist(tempoAuto, 0, _maxTicks, w.PlaylistTracks[200]); // TODO: track
-
-			tempoAuto.AddTempoPoint((uint)firstTempo.Ticks, tempo);
-		}
-		tempoAuto.AddTempoPoint((uint)e.Ticks, bpm);
-	}
-	private static void FLP_HandleTimeSig(MIDIEvent e, MetaMessage m, FLProjectWriter w,
-		ref byte timeSigNum, ref byte timeSigDenom, ref MIDIEvent? firstTimeSig, ref bool createdFirstTimeSigMarker)
-	{
-		m.ReadTimeSignatureMessage(out byte num, out byte denom, out _, out _);
-		if (firstTimeSig is null)
-		{
-			timeSigNum = num;
-			timeSigDenom = denom;
-			firstTimeSig = e;
-			return;
-		}
-
-		// This is the 2nd or after change. 2nd will create the marker for #1 and #2
-		if (!createdFirstTimeSigMarker)
-		{
-			createdFirstTimeSigMarker = true;
-			w.AddTimeSigMarker((uint)firstTimeSig.Ticks, timeSigNum, timeSigDenom);
-		}
-		w.AddTimeSigMarker((uint)e.Ticks, num, denom);
 	}
 }
